Split utils and fixed types (#1030)
* split utils

* fixed types
