+++ autokeras/__init__.py
@@ -33,7 +33,6 @@ from autokeras.tasks import StructuredDataClassifier
 from autokeras.tasks import StructuredDataRegressor
 from autokeras.tasks import TextClassifier
 from autokeras.tasks import TextRegressor
-
-from .utils import check_tf_version
+from autokeras.utils.utils import check_tf_version
 
 check_tf_version()
