+++ autokeras/adapters/output_adapter.py
@@ -3,8 +3,8 @@ import pandas as pd
 import tensorflow as tf
 
 from autokeras import encoders
-from autokeras import utils
 from autokeras.engine import adapter as adapter_module
+from autokeras.utils import data_utils
 
 
 class HeadAdapter(adapter_module.Adapter):
@@ -65,7 +65,7 @@ class ClassificationHeadAdapter(HeadAdapter):
         # If in tf.data.Dataset, must be encoded already.
         if isinstance(dataset, tf.data.Dataset):
             if not self.num_classes:
-                shape = utils.dataset_shape(dataset)[0]
+                shape = data_utils.dataset_shape(dataset)[0]
                 # Single column with 0s and 1s.
                 if shape == 1:
                     self.num_classes = 2
