+++ autokeras/tasks/structured_data.py
@@ -9,8 +9,8 @@ import pandas as pd
 from autokeras import auto_model
 from autokeras import hypermodels
 from autokeras import nodes as input_module
-from autokeras import utils
 from autokeras.tuners import greedy
+from autokeras.utils import types
 
 
 class SupervisedStructuredDataPipeline(auto_model.AutoModel):
@@ -295,8 +295,8 @@ class StructuredDataRegressor(SupervisedStructuredDataPipeline):
                  column_names: Optional[List[str]] = None,
                  column_types: Optional[Dict[str, str]] = None,
                  output_dim: Optional[int] = None,
-                 loss: utils.AcceptableLoss = 'mean_squared_error',
-                 metrics: utils.AcceptableMetrics = None,
+                 loss: types.LossType = 'mean_squared_error',
+                 metrics: Optional[types.MetricsType] = None,
                  name: str = 'structured_data_regressor',
                  max_trials: int = 100,
                  directory: Union[str, pathlib.Path, None] = None,
