+++ autokeras/tasks/text.py
@@ -5,9 +5,9 @@ from typing import Union
 from autokeras import auto_model
 from autokeras import hypermodels
 from autokeras import nodes as input_module
-from autokeras import utils
 from autokeras.tuners import greedy
 from autokeras.tuners import task_specific
+from autokeras.utils import types
 
 
 class SupervisedTextPipeline(auto_model.AutoModel):
@@ -45,8 +45,8 @@ class TextClassifier(SupervisedTextPipeline):
     def __init__(self,
                  num_classes: Optional[int] = None,
                  multi_label: bool = False,
-                 loss: utils.AcceptableLoss = None,
-                 metrics: utils.AcceptableMetrics = None,
+                 loss: types.LossType = None,
+                 metrics: Optional[types.MetricsType] = None,
                  name: str = 'text_classifier',
                  max_trials: int = 100,
                  directory: Union[str, pathlib.Path, None] = None,
