+++ autokeras/utils/layer_utils.py
@@ -1,0 +1,37 @@
+import tensorflow as tf
+
+
+def get_global_average_pooling(shape):
+    return [tf.keras.layers.GlobalAveragePooling1D,
+            tf.keras.layers.GlobalAveragePooling2D,
+            tf.keras.layers.GlobalAveragePooling3D][len(shape) - 3]
+
+
+def get_global_max_pooling(shape):
+    return [tf.keras.layers.GlobalMaxPool1D,
+            tf.keras.layers.GlobalMaxPool2D,
+            tf.keras.layers.GlobalMaxPool3D][len(shape) - 3]
+
+
+def get_max_pooling(shape):
+    return [tf.keras.layers.MaxPool1D,
+            tf.keras.layers.MaxPool2D,
+            tf.keras.layers.MaxPool3D][len(shape) - 3]
+
+
+def get_conv(shape):
+    return [tf.keras.layers.Conv1D,
+            tf.keras.layers.Conv2D,
+            tf.keras.layers.Conv3D][len(shape) - 3]
+
+
+def get_sep_conv(shape):
+    return [tf.keras.layers.SeparableConv1D,
+            tf.keras.layers.SeparableConv2D,
+            tf.keras.layers.Conv3D][len(shape) - 3]
+
+
+def get_dropout(shape):
+    return [tf.keras.layers.SpatialDropout1D,
+            tf.keras.layers.SpatialDropout2D,
+            tf.keras.layers.SpatialDropout3D][len(shape) - 3]
